requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# Machine Learning & Data Analysis
scikit-learn==1.3.2
XGBoost==2.0.1
pandas==2.1.4
numpy==1.25.2
ta==0.10.2

# Financial Data
yfinance==0.2.28
requests==2.31.0

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Date/Time Handling
python-dateutil==2.8.2

# HTTP Requests
urllib3==2.1.0
certifi==2023.11.17

# Optional: For production deployment
gunicorn==21.2.0